refactor: extract recaptcha hook (temp site)
Made-with: Cursor
